requirements.txt
altair==4.2.0
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
filelock==3.8.0
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
htbuilder==0.6.1
huggingface-hub==0.10.1
idna==3.4        
importlib-metadata==5.0.0
importlib-resources==5.10.0
Jinja2==.1.2
jsonschema==4.17.0
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
more-itertools==9.0.0
numpy==1.21.6     
packaging==21.3
pandas==1.3.5
Pillow==9.3.0
pkgutil-resolve-name==1.3.10
plotly==5.11.0     
protobuf==3.20.3
psycopg2==2.9.5
pyarrow==10.0.0
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31 
requests==2.28.1
rich==12.6.0
scipy==1.7.3
semver==2.13.0
setuptools==41.2.0
six==1.16.0
smmap==5.0.0
snscrape==0.3.4      
st-annotated-text==3.0.0
streamlit==1.14.1
tenacity==8.1.0
tokenizers==0.13.2     
torch==1.13.0
transformers==4.24.0